requirements.txt
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.1.0
ipykernel==6.29.5
ipytest==0.14.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nbclient==0.10.2
nbformat==5.10.4
nbmake==1.5.5
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
referencing==0.36.2
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
python-dotenv==1.0.0